requirements.txt
about-time==4.2.1
absl-py==1.4.0
alive-progress==3.1.4
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
colorcet==3.0.1
contourpy==1.0.7
cycler==0.11.0
dask==2023.6.0
datashader==0.15.0
datashape==0.5.2
Django==4.2.2
docx==0.2.4
flatbuffers==23.5.26
fonttools==4.39.4
fsspec==2023.6.0
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grapheme==0.6.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
image==1.5.33
imageio==2.31.0
importlib-metadata==6.6.0
jax==0.4.12
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
llvmlite==0.40.1rc1
locket==1.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-dtypes==0.2.0
multipledispatch==0.6.0
networkx==3.1
numba==0.57.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
pandas-io==0.0.1
param==1.13.0
partd==1.4.0
Pillow==9.4.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyct==0.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.14.1
xarray==2023.5.0
zipp==3.15.0
